+++ mod.ts
@@ -8,3 +8,4 @@ export { map } from "./map.ts";
 export { take, takeWhile } from "./take.ts";
 export { tee } from "./tee.ts";
 export { assertStreams, assertStreamStartsWith } from "./testing.ts";
+export { unique } from "./unique.ts";